refactor: extract model and profile domain types
