+++ src/domain.rs
@@ -1,5 +1,7 @@
 pub(crate) mod activity;
 pub(crate) mod chat;
 pub(crate) mod elicitation;
+pub(crate) mod model;
+pub(crate) mod profile;
 pub(crate) mod session;
 pub(crate) mod tool;
